requirements.txt
boto==2.39.0
contextlib2==0.5.1
Django==1.9.4
ecdsa==0.13
Fabric==1.10.2
paramiko==1.16.0
psycopg2==2.6.1
pycrypto==2.6.1
raven==5.11.1
wheel==0.26.0
django-filter==0.13.0
django-storages-redux==1.3.2
django-wysiwyg-redactor==0.4.9
djangorestframework==3.3.3
Markdown==2.6.5
